requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.0
rich>=13.7.0
Pillow>=10.0.0
